Create controllers for Categories and Cart + endpoints according to CRUD

+++ Scala/app/models/Product.scala
@@ -3,7 +3,7 @@ package models
 import play.api.libs.json._
 
 case class Product (
-    id: Option[Long], 
+    id: Option[Int], 
     title: String, 
     description: Option[String], 
     price: BigDecimal
